LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/MediaQueryList-change-event-matches-value-expected.txt


PASS MediaQueryList.changed is correct for all lists in the document even during a change event handler
